requirements.txt
aiofiles==0.6.0
aiohttp==3.8.4
aiohttp-socks==0.7.1
aiosignal==1.3.1
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.2.0
blobfile==2.0.1
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
jsonschema==4.17.3
Logbook==1.5.3
lxml==4.9.2
matrix-nio==0.20.1
multidict==6.0.4
peewee==3.16.0
Pillow==9.4.0
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pyrsistent==0.19.3
python-magic==0.4.27
python-olm==3.1.3
python-socks==2.1.1
regex==2022.10.31
requests==2.28.2
tiktoken==0.3.0
unpaddedbase64==2.1.0
urllib3==1.26.14
wcwidth==0.2.6
yarl==1.8.2
